Change to positioning.
With the positioning set to absolute there are adverse affects to the surrounding page, especially if used within a modal, due to stretching. As it's already inside a container (<div class="ng-ms form-item-container">)setting position:relative to it, allows it to stretch and float above while still conforming to the surrounding structure.