Added a max height and overflow scroll.
When the select list extends beyond 5 or so items it becomes too long to sensibly fit on the page. Have added a <b>max-height of 300px</b> as a more sensible maximum, and added <b>overflow: auto</b> in order to add a scroll bar IF the content extends beyond the specified pixels.